zh_CN Locales Eye Of The Queen
Eye Of The Queen (ID:220003) is a enemy in the RP event before the second boss in the City of Threads.
Before being discovered, this monster uses the name "Hollows Resident" to hide enemy numerous NPCs.
When the player discovers and exposes it as a spy, it transforms into Eye Of The Queen.
(This is similar to the spy-finding storyline in Court of Stars.)

Eye Of The Queen's Chinese name is "女王之眼"
Hollows Resident's Chinese name is "高耸虚居居民"

Therefore, in both the naming of enemy in CityofThreads.lua and the zh_CN locales translation, I use "Eye Of The Queen"(女王之眼)